+++ packages/manager/src/store/authentication/authentication.test.ts
@@ -1,10 +1,13 @@
-import store from 'src/store';
+import { queryClientFactory as queryClientFactory } from 'src/queries/base';
 import { authentication } from 'src/utilities/storage';
+import { storeFactory } from '..';
 import {
   handleInitTokens,
   handleLogout,
   handleStartSession,
 } from './authentication.actions';
+
+const store = storeFactory(queryClientFactory());
 
 describe('Authentication', () => {
   authentication.expire.set('hello world');