+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Fixed:
 - Typesafety of the `<Select />` component #8986
 
+### Tech Stories:
+- MUIv5 Migration - Components > TagsInput, TagsPanel #8995
+
 ## [2023-04-03] - v1.90.0
 
 ### Added: